Changed size_t to uint_64 to ensure that we support 64-bit sizes even on 32-bit systems

+++ hash_extender_sha256.c
@@ -13,11 +13,11 @@
 #define SHA256_BLOCK 64
 
 /* Note: this only supports data with a 4-byte size (4.2 billion bits). */
-uint8_t *sha256_append_data(uint8_t *data, size_t data_length, size_t secret_length, uint8_t *append, size_t append_length, size_t *new_length)
+uint8_t *sha256_append_data(uint8_t *data, uint64_t data_length, size_t secret_length, uint8_t *append, size_t append_length, size_t *new_length)
 {
   /* Allocate memory for the new buffer (enough room for buffer + a full block + the data) */
   uint8_t *result = (uint8_t*) malloc(1000 + data_length + append_length + SHA256_BLOCK); /* (This can overflow if we're ever using this in a security-sensitive context) */
-  size_t bit_length;
+  uint64_t bit_length;
 
   /* Start with the current buffer and length. */
   memmove(result, data, data_length);
@@ -47,7 +47,7 @@ uint8_t *sha256_append_data(uint8_t *data, size_t data_length, size_t secret_len
   return result;
 }
 
-void sha256_gen_signature(uint8_t *secret, size_t secret_length, uint8_t *data, size_t data_length, uint8_t signature[SHA256_DIGEST_LENGTH])
+void sha256_gen_signature(uint8_t *secret, uint64_t secret_length, uint8_t *data, size_t data_length, uint8_t signature[SHA256_DIGEST_LENGTH])
 {
   SHA256_CTX c;
   SHA256_Init(&c);
@@ -56,11 +56,11 @@ void sha256_gen_signature(uint8_t *secret, size_t secret_length, uint8_t *data, 
   SHA256_Final(signature, &c);
 }
 
-void sha256_gen_signature_evil(size_t secret_length, size_t data_length, uint8_t original_signature[SHA256_DIGEST_LENGTH], uint8_t *append, size_t append_length, uint8_t new_signature[SHA256_DIGEST_LENGTH])
+void sha256_gen_signature_evil(uint64_t secret_length, size_t data_length, uint8_t original_signature[SHA256_DIGEST_LENGTH], uint8_t *append, size_t append_length, uint8_t new_signature[SHA256_DIGEST_LENGTH])
 {
   SHA256_CTX c;
-  size_t original_data_length;
-  size_t i;
+  uint64_t original_data_length;
+  uint64_t i;
 
   SHA256_Init(&c);
 
@@ -88,7 +88,7 @@ void sha256_gen_signature_evil(size_t secret_length, size_t data_length, uint8_t
   SHA256_Final(new_signature, &c);
 }
 
-static int sha256_test_validate(uint8_t *secret, size_t secret_length, uint8_t *data, size_t data_length, uint8_t *signature)
+static int sha256_test_validate(uint8_t *secret, uint64_t secret_length, uint8_t *data, size_t data_length, uint8_t *signature)
 {
   unsigned char result[SHA256_DIGEST_LENGTH];
 
@@ -107,7 +107,7 @@ static void sha256_test_extension()
   uint8_t *data      = (uint8_t*)"DATA";
   uint8_t *append    = (uint8_t*)"APPEND";
   uint8_t *new_data;
-  size_t  new_length;
+  uint64_t  new_length;
 
   uint8_t original_signature[SHA256_DIGEST_LENGTH];
   uint8_t new_signature[SHA256_DIGEST_LENGTH];
@@ -135,12 +135,12 @@ static void sha256_test_lengths()
   uint8_t *data      = (uint8_t*)"BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB";
   uint8_t *append    = (uint8_t*)"CCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCC";
   uint8_t *new_data;
-  size_t  new_length;
+  uint64_t  new_length;
 
   uint8_t original_signature[SHA256_DIGEST_LENGTH];
   uint8_t new_signature[SHA256_DIGEST_LENGTH];
 
-  size_t i;
+  uint64_t i;
 
   printf("Testing SHA256 data of various lengths...\n");
