+++ hash_extender_sha512.c
@@ -11,11 +11,11 @@
 
 #define SHA512_BLOCK 64
 
-uint8_t *sha512_append_data(uint8_t *data, size_t data_length, size_t secret_length, uint8_t *append, size_t append_length, size_t *new_length)
+uint8_t *sha512_append_data(uint8_t *data, uint64_t data_length, size_t secret_length, uint8_t *append, size_t append_length, size_t *new_length)
 {
   /* Allocate memory for the new buffer (enough room for buffer + a full block + the data) */
   uint8_t *result = (uint8_t*) malloc(1000 + data_length + append_length + SHA512_BLOCK); /* (This can overflow if we're ever using this in a security-sensitive context) */
-  size_t bit_length;
+  uint64_t bit_length;
 
   /* Start with the current buffer and length. */
   memmove(result, data, data_length);
@@ -45,7 +45,7 @@ uint8_t *sha512_append_data(uint8_t *data, size_t data_length, size_t secret_len
   return result;
 }
 
-void sha512_gen_signature(uint8_t *secret, size_t secret_length, uint8_t *data, size_t data_length, uint8_t signature[SHA512_DIGEST_LENGTH])
+void sha512_gen_signature(uint8_t *secret, uint64_t secret_length, uint8_t *data, size_t data_length, uint8_t signature[SHA512_DIGEST_LENGTH])
 {
   SHA512_CTX c;
   SHA512_Init(&c);
@@ -54,11 +54,11 @@ void sha512_gen_signature(uint8_t *secret, size_t secret_length, uint8_t *data, 
   SHA512_Final(signature, &c);
 }
 
-void sha512_gen_signature_evil(size_t secret_length, size_t data_length, uint8_t original_signature[SHA512_DIGEST_LENGTH], uint8_t *append, size_t append_length, uint8_t new_signature[SHA512_DIGEST_LENGTH])
+void sha512_gen_signature_evil(uint64_t secret_length, size_t data_length, uint8_t original_signature[SHA512_DIGEST_LENGTH], uint8_t *append, size_t append_length, uint8_t new_signature[SHA512_DIGEST_LENGTH])
 {
   SHA512_CTX c;
-  size_t original_data_length;
-  size_t i;
+  uint64_t original_data_length;
+  uint64_t i;
 
   SHA512_Init(&c);
 
@@ -86,7 +86,7 @@ void sha512_gen_signature_evil(size_t secret_length, size_t data_length, uint8_t
   SHA512_Final(new_signature, &c);
 }
 
-static int sha512_test_validate(uint8_t *secret, size_t secret_length, uint8_t *data, size_t data_length, uint8_t *signature)
+static int sha512_test_validate(uint8_t *secret, uint64_t secret_length, uint8_t *data, size_t data_length, uint8_t *signature)
 {
   unsigned char result[SHA512_DIGEST_LENGTH];
 
@@ -105,7 +105,7 @@ static void sha512_test_extension()
   uint8_t *data      = (uint8_t*)"DATA";
   uint8_t *append    = (uint8_t*)"APPEND";
   uint8_t *new_data;
-  size_t  new_length;
+  uint64_t  new_length;
 
   uint8_t original_signature[SHA512_DIGEST_LENGTH];
   uint8_t new_signature[SHA512_DIGEST_LENGTH];
@@ -133,12 +133,12 @@ static void sha512_test_lengths()
   uint8_t *data      = (uint8_t*)"BBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBBB";
   uint8_t *append    = (uint8_t*)"CCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCCC";
   uint8_t *new_data;
-  size_t  new_length;
+  uint64_t  new_length;
 
   uint8_t original_signature[SHA512_DIGEST_LENGTH];
   uint8_t new_signature[SHA512_DIGEST_LENGTH];
 
-  size_t i;
+  uint64_t i;
 
   printf("Testing SHA512 data of various lengths...\n");
