+++ util.h
@@ -7,6 +7,6 @@
 #include <stdint.h>
 
 void print_hex(unsigned char *data, unsigned int length);
-void print_hex_fancy(uint8_t *data, size_t length);
+void print_hex_fancy(uint8_t *data, uint64_t length);
 
 #endif
